requirements.txt
fastapi
uvicorn
Pillow
google-cloud-vision
geopy
python-multipart
